Non-funcitonal Changes made to the html / php.
Non-funcitoning update recording several changes
to the css for the mobile first redesign. HTML
and php funcitons, but the js is broken.
Also, several changes were made to html elements
The names, ids, and classes of many elements were
also changed, to be more standards compliant.
php was updated for modularity.